Extract GridViewPanel
